+++ runtime/core/xnix_runtime_core_cli.c
@@ -2822,6 +2822,8 @@ print_probe(void)
   fputs("\"desktop_entry_plan_count\":1,", stdout);
   fputs("\"task_manager_identity_plan_owner\":\"c\",", stdout);
   fputs("\"task_manager_identity_plan_count\":1,", stdout);
+  fputs("\"file_association_plan_owner\":\"c\",", stdout);
+  fputs("\"file_association_plan_count\":1,", stdout);
   fputs("\"portal_request_plan_owner\":\"c\",", stdout);
   fputs("\"portal_request_plan_count\":7,", stdout);
   fputs("\"write_methods\":", stdout);
@@ -2880,5 +2882,6 @@ print_write_gate(const char *method_name)
 #include "xnix_runtime_core_cli_desktop_activation.inc"
 #include "xnix_runtime_core_cli_desktop_entry.inc"
 #include "xnix_runtime_core_cli_task_manager_identity.inc"
+#include "xnix_runtime_core_cli_file_association.inc"
 #include "xnix_runtime_core_cli_portal_request.inc"
 #include "xnix_runtime_core_cli_dispatch.inc"